'config.php' file moved one level up
